makeconf_path {rstan}	R Documentation

Obtain the full path of file Makeconf

Description

Obtain the full path of file Makeconf, in which, for example the flags for compiling C/C++ code are configured.

Usage

makeconf_path() 

Details

The configuration for compiling shared objects using R CMD SHLIB are set in file Makeconf. To change how the C++ code is compiled, modify this file. For RStan, package inline compiles the C++ code using R CMD SHLIB. To speed up compiled Stan models, increase the optimization level to -O3 defined in property CXXFLAGS in the file Makeconf. This file may also be modified to specify alternative C++ compilers, such as clang++ or later versions of g++.

Value

An character string for the full path of file Makeconf.

See Also

stan

Examples

makeconf_path() 

[Package rstan version 2.17.2 Index]