rstan-plotting-functions {rstan}	R Documentation

RStan Plotting Functions

Description

List of RStan plotting functions that return ggplot objects

RStan plotting functions

Posterior intervals and point estimates

stan_plot

Traceplots

stan_trace

Histograms

stan_hist

Kernel density estimates

stan_dens

Scatterplots

stan_scat

Diagnostics for Hamiltonian Monte Carlo and the No-U-Turn Sampler

stan_diag

Rhat

stan_rhat

Ratio of effective sample size to total posterior sample size

stan_ess

Ratio of Monte Carlo standard error to posterior standard deviation

stan_mcse

Autocorrelation

stan_ac

See Also

Plot options

[Package rstan version 2.17.2 Index]