read_rdump {rstan}	R Documentation

Read data in an R dump file to a list

Description

Create an R list from an R dump file

Usage

read_rdump(f, keep.source = FALSE, ...)

Arguments

f
A character string providing the dump file name.

keep.source
logical: should the source formatting be retained when echoing expressions, if possible?

...
passed to source

Details

The R dump file can be read directly by R function source, which by default would read the data into the user's workspace (the global environment). This function instead read the data to a list, making it convenient to prepare data for the stan model-fitting function.

Value

A list containing all the data defined in the dump file with keys corresponding to variable names.

See Also

stan_rdump; dump

Examples

x <- 1; y <- 1:10; z <- array(1:10, dim = c(2,5)) 
stan_rdump(ls(pattern = '^[xyz]'), "xyz.Rdump")
l <- read_rdump('xyz.Rdump')
print(l) 

[Package rstan version 2.17.2 Index]