rstan-internal {rstan}	R Documentation

Internal Functions and Methods

Description

These functions are not intended to be called directly. In the case of methods, they are documented elsewhere, either with the associated generic function or with the class of the object for which the method is defined.

[Package rstan version 2.17.2 Index]