set_cppo {rstan}	R Documentation

Defunct function to set the compiler optimization level

Description

This function returns nothing and does nothing except throw a warning. See https://cran.r-project.org/doc/manuals/r-release/R-admin.html#Customizing-package-compilation for information on customizing the compiler options, but doing so should be unnecessary for normal useage.

Usage

  set_cppo(...)

Arguments

...
Any input is ignored

Value

An invisible NULL

[Package rstan version 2.17.2 Index]