stan_version {rstan}	R Documentation

Obtain the version of Stan

Description

The stan version is in form of major.minor.patch; the first version is 1.0.0, indicating major version 1, minor version 0 and patch level 0. Functionality only changes with minor versions and backward compatibility will only be affected by major versions.

Usage

  stan_version()

Value

A character string giving the version of Stan used in this version of RStan.

References

The Stan Development Team Stan Modeling Language User's Guide and Reference Manual. http://mc-stan.org/.

See Also

stan and stan_model

Examples

  stan_version() 

[Package rstan version 2.17.2 Index]