Look up the Stan function that corresponds to a R function or name.

This function helps to map between R functions and Stan functions.

lookup(FUN, ReturnType = character())

Arguments

FUN
A character string naming a R function or a R function for which the (near) equivalent Stan function is sought. If no matching R function is found, FUN is reinterpreted as a regexp and matches are sought.

ReturnType
A character string of positive length naming a valid return type for a Stan function: int, int[], matrix, real, real[,], real[], row_vector, T[], vector, or void. If "ANY" is passed, then the entire data.frame is returned and can be inspected with the View function, for example.

Value

Ordinarily, a data.frame with rows equal to the number of partial matches and four columns:

StanFunction Character string for the Stan function's name.

Arguments Character string indicating the arguments to that Stan function.

ReturnType Character string indicating the return type of that Stan function.

Page Integer indicating the page of the Stan reference manual where that Stan function is defined.

If there are no matching Stan functions, a character string indicating so is returned.

References

The Stan Development Team Stan Modeling Language User's Guide and Reference Manual. http://mc-stan.org/.

The Stan Development Team CmdStan Interface User's Guide. http://mc-stan.org.

Examples

lookup(dnorm)      # Stan equivalents for the normal PDF (in log form)

#>     StanFunction                         Arguments ReturnType Page
#> 353  normal_lpdf (reals y | reals mu, reals sigma)       real  493
#> 354       normal                                 ~       real  493

lookup("foo")      # fails

#> [1] "no matching Stan functions"

lookup("Student")  # succeeds even though there is no such R function

#>             StanFunction                                              Arguments
#> 328 multi_student_t_lpdf (row_vectors y | real nu, row_vectors mu matrix Sigma)
#> 329 multi_student_t_lpdf     (row_vectors y | real nu, vectors mu matrix Sigma)
#> 330 multi_student_t_lpdf     (vectors y | real nu, row_vectors mu matrix Sigma)
#> 331 multi_student_t_lpdf         (vectors y | real nu, vectors mu matrix Sigma)
#> 332      multi_student_t                                                      ~
#> 333  multi_student_t_rng                     (real nu, vector mu, matrix Sigma)
#> 569        student_t_cdf             (reals y, reals nu, reals mu, reals sigma)
#> 570      student_t_lccdf            (reals y | reals nu, reals mu, reals sigma)
#> 571       student_t_lcdf            (reals y | reals nu, reals mu, reals sigma)
#> 572       student_t_lpdf            (reals y | reals nu, reals mu, reals sigma)
#> 573            student_t                                                      ~
#> 574        student_t_rng                         (real nu, real mu, real sigma)
#>     ReturnType Page
#> 328       real  523
#> 329       real  523
#> 330       real  523
#> 331       real  523
#> 332       real  522
#> 333     vector  523
#> 569       real  496
#> 570       real  496
#> 571       real  496
#> 572       real  496
#> 573       real  496
#> 574       real  496

lookup("^poisson") # every Stan function that starts with poisson

#>         StanFunction               Arguments ReturnType Page
#> 466      poisson_cdf  (ints n, reals lambda)       real  489
#> 467    poisson_lccdf (ints n | reals lambda)       real  489
#> 468     poisson_lcdf (ints n | reals lambda)       real  489
#> 469 poisson_log_lpmf  (ints n | reals alpha)       real  490
#> 470      poisson_log                       ~       real  490
#> 471  poisson_log_rng            (real alpha)        int  490
#> 472     poisson_lpmf (ints n | reals lambda)       real  489
#> 473          poisson                       ~       real  489
#> 474      poisson_rng           (real lambda)        int  489