Look up the Stan function that corresponds to a R function or name.

This function helps to map between R functions and Stan functions.

lookup(FUN, ReturnType = character())

Arguments

FUN
A character string naming a R function or a R function for which the (near) equivalent Stan function is sought. If no matching R function is found, FUN is reinterpreted as a regexp and matches are sought.

ReturnType
A character string of positive length naming a valid return type for a Stan function: int, int[], matrix, real, real[,], real[], row_vector, T[], vector, or void. If "ANY" is passed, then the entire data.frame is returned and can be inspected with the View function, for example.

Value

Ordinarily, a data.frame with rows equal to the number of partial matches and four columns:

StanFunction Character string for the Stan function's name.

Arguments Character string indicating the arguments to that Stan function.

ReturnType Character string indicating the return type of that Stan function.

Page Integer indicating the page of the Stan reference manual where that Stan function is defined.

If there are no matching Stan functions, a character string indicating so is returned.

References

The Stan Development Team Stan Modeling Language User's Guide and Reference Manual. http://mc-stan.org/.

The Stan Development Team CmdStan Interface User's Guide. http://mc-stan.org.

Examples

lookup(dnorm)      # Stan equivalents for the normal PDF (in log form)

#>     StanFunction                         Arguments ReturnType Page
#> 355  normal_lpdf (reals y | reals mu, reals sigma)       real  505
#> 356       normal                                 ~       real  505

lookup("foo")      # fails

#> [1] "no matching Stan functions"

lookup("Student")  # succeeds even though there is no such R function

#>             StanFunction                                              Arguments
#> 330 multi_student_t_lpdf (row_vectors y | real nu, row_vectors mu matrix Sigma)
#> 331 multi_student_t_lpdf     (row_vectors y | real nu, vectors mu matrix Sigma)
#> 332 multi_student_t_lpdf     (vectors y | real nu, row_vectors mu matrix Sigma)
#> 333 multi_student_t_lpdf         (vectors y | real nu, vectors mu matrix Sigma)
#> 334      multi_student_t                                                      ~
#> 335  multi_student_t_rng                     (real nu, vector mu, matrix Sigma)
#> 608        student_t_cdf             (reals y, reals nu, reals mu, reals sigma)
#> 609      student_t_lccdf            (reals y | reals nu, reals mu, reals sigma)
#> 610       student_t_lcdf            (reals y | reals nu, reals mu, reals sigma)
#> 611       student_t_lpdf            (reals y | reals nu, reals mu, reals sigma)
#> 612            student_t                                                      ~
#> 613        student_t_rng                         (real nu, real mu, real sigma)
#>     ReturnType Page
#> 330       real  535
#> 331       real  535
#> 332       real  535
#> 333       real  535
#> 334       real  534
#> 335     vector  535
#> 608       real  508
#> 609       real  508
#> 610       real  508
#> 611       real  508
#> 612       real  508
#> 613       real  508

lookup("^poisson") # every Stan function that starts with poisson

#>         StanFunction               Arguments ReturnType Page
#> 505      poisson_cdf  (ints n, reals lambda)       real  501
#> 506    poisson_lccdf (ints n | reals lambda)       real  501
#> 507     poisson_lcdf (ints n | reals lambda)       real  501
#> 508 poisson_log_lpmf  (ints n | reals alpha)       real  502
#> 509      poisson_log                       ~       real  502
#> 510  poisson_log_rng            (real alpha)        int  502
#> 511     poisson_lpmf (ints n | reals lambda)       real  501
#> 512          poisson                       ~       real  501
#> 513      poisson_rng           (real lambda)        int  501