RStan --- R interface to Stan

Stan Development Team

RStan is the R interface to the Stan C++ package. RStan provides

full Bayesian inference using the No-U-Turn sampler (NUTS), a variant of Hamiltonian Monte Carlo (HMC)

approximate Bayesian inference using automatic differentiation variational inference (ADVI)

penalized maximum likelihood estimation using L-BFGS optimization

For documentation on the Stan modeling language see the Stan Modeling Language User's Guide and Reference Manual.

Other R packages from the Stan Development Team

Various related R packages are also available from the Stan Development Team:

Package 	 Description 	 Link

bayesplot 	 ggplot-based plotting library for graphing parameter estimates, MCMC diagnostics, and posterior predictive checks. 	 bayesplot-package

shinystan 	 Interactive visual summaries and advanced posterior analysis of MCMC output. 	 shinystan-package

loo 	 Out-of-sample predictive performance estimates and model comparison. 	 loo-package

rstanarm 	 R formula interface for Bayesian applied regression modeling. 	 rstanarm-package

rstantools 	 Tools for developers of R packages interfacing with Stan. 	 rstantools-package

See also

stan for details on fitting models and stanfit for information on the fitted model objects.

https://github.com/stan-dev/rstan/issues/ to submit a bug report or feature request.

http://discourse.mc-stan.org to ask a question on the Stan-users forum.

Examples